fix: resolve conflicts after rebase with main

+++ db/sql/migrations/20251003112638_create_table_order_items.sql
@@ -7,9 +7,9 @@ CREATE TABLE order_items (
     item_name VARCHAR(255),
     product_category VARCHAR(255),
     item_image_url VARCHAR(255),
-    price_per_item INTEGER NOT NULL,
+    price_per_item NUMERIC(12,2) NOT NULL,
     quantity INTEGER NOT NULL,
-    total_price INTEGER NOT NULL,
+    total_price NUMERIC(12,2) NOT NULL,
     created_at TIMESTAMP DEFAULT NOW(),
     updated_at TIMESTAMP DEFAULT NOW(),
     CONSTRAINT order_items_quantity_positive CHECK (quantity > 0)
